[refactor]: 优化Board和piece的逻辑, 对 pos 等属性进行封装
1. 优化 Board 和 Piece 的逻辑
2. 拆分 Piece.ts 为三个文件
3. 添加文件夹整理各个文件
