+++ src/Game/GetData.ts
@@ -1,8 +1,9 @@
 import {Board, Grid} from "./Board";
-import {Piece, PieceStatic} from "./Piece";
+import {Piece} from "./Piece/Piece";
+import {PieceStatic} from "./Piece/PieceStatic";
 import {Game} from "./Game";
 import {Pos} from "./Pos";
-import {Player} from "./Player";
+import {Player} from "./Player/Player";
 
 export class BoardData {
     size: Pos;
