+++ src/Game/Player/PlayerManager.ts
@@ -1,6 +1,6 @@
 import {Player} from "./Player";
 import {JSX} from "react";
-import {Game} from "./Game";
+import {Game} from "../Game";
 
 export class PlayerManager {
