+++ src/Game/Rules/Rules.ts
@@ -1,6 +1,6 @@
-import {Board} from "./Board";
-import {Piece, PieceStatic} from "./Piece";
-import {Game} from "./Game";
+import {Board} from "../Board";
+import {Piece, PieceStatic} from "../Piece/Piece";
+import {Game} from "../Game";
 
 
 //每种棋子都有自己的规则, 这些规则会放在本 class 的子类中
